This is a profile for PrusaSlicer.
Here is a basic profile that I created for the Fokoos Odin-5 F3 Printer for PrusaSlicer.  It was created in late 2021 and I have not had an Odin-5 for a while so I have not been able to tweak and continue working on it but it is a great base file to start your tweaking with.